% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFTE_epoch_tbl.R
\name{SAFTE_epoch_tbl}
\alias{SAFTE_epoch_tbl}
\title{SAFTE Epoch Table}
\usage{
SAFTE_epoch_tbl(
  dataset,
  subject_id,
  obs_start,
  obs_end,
  sleep_id = "Sleep",
  work_id = "Work",
  crewing_id = "Crewing",
  test_id = "Test"
)
}
\arguments{
\item{dataset}{Name of the properly formatted dataset uploaded into the R environment by the user (Required)}

\item{subject_id}{Identifier of the subject/participant/patient. (Required)}

\item{obs_start}{Oberservation start time (i.e. beginning of the study). Formatted as 'YYYY-mm-dd HH:MM' (Required)}

\item{obs_end}{Observation end time (i.e. end of the study). Cannot be more than 21 days from the obs_start. Formatted as 'YYYY-mm-dd HH:MM' (Required)}

\item{sleep_id}{Marker ID for a sleep period (usually "Sleep" or "sleep"). Defaults to "Sleep"}

\item{work_id}{Marker ID for a work period. Defaults to "Work"}

\item{crewing_id}{Marker ID for a crewing period within a work period. Defaults to "Crewing"}

\item{test_id}{Marker ID for a test period (i.e. PVT taken or medication given). Defaults to "Test"}
}
\value{
A list 4 tables. One for identified periods of sleep, work, testing, and/or crewing if applicable. Period types not identified will return a table as Null.
}
\description{
This function takes the raw, properly formatted data uploaded by the user and
creates tables of the 1 minute Epoch data. This is the first essential step
for using the SAFTE model.
}
\details{
Properly formatted data at a minimum has a single column for
sleep identification, a single column for start datetimes for bedtime, and a
single column for end dateimes for waketimes. It can also include work, test,
and crewing identified periods within the same columns.

This is not for EPOCH by EPOCH data or data already divided into 1 minute
epochs.
}
