% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFTE_missing_sleep.R
\name{SAFTE_missing_sleep}
\alias{SAFTE_missing_sleep}
\title{SAFTE Missing Sleep}
\usage{
SAFTE_missing_sleep(
  dataset,
  subject_id,
  split_time = 12,
  sleep_id,
  obs_start,
  obs_end
)
}
\arguments{
\item{dataset}{Name of the properly formatted dataset uploaded into the R environment by the user (Required)}

\item{subject_id}{Identifier of the subject/participant/patient. (Required)}

\item{split_time}{#the hour (0-23) at which the user wants to split the day. Deafults to noon ("12"). (Required)}

\item{sleep_id}{Marker ID for a sleep period (usually "Sleep" or "sleep"). Defaults to "Sleep"#'}

\item{obs_start}{Oberservation start time (i.e. beginning of the study). Formatted as 'YYYY-mm-dd HH:MM' (Required)}

\item{obs_end}{Observation end time (i.e. end of the study). Cannot be more than 21 days from the obs_start. Formatted as 'YYYY-mm-dd HH:MM' (Required)}
}
\value{
This will output a table listing all 24 hour periods that do not have an instance of sleep.
}
\description{
This function takes the raw, properly formatted data uploaded by the user and
searches for any dates that do not include "Sleep". This could be helpful in
identifying missing night of sleep data which can dramatically affect the
calculations in the SAFTE model.
}
\details{
Properly formatted data at a minimum has a single column for
sleep identification, a single column for start datetimes for bedtime, and a
single column for end dateimes for waketimes. It can also include work, test,
and crewing identified periods within the same columns.

This is not for EPOCH by EPOCH data or data already divided into 1 minute
epochs.
}
