% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFTE_model.R
\name{SAFTE_model}
\alias{SAFTE_model}
\title{SAFTE Model}
\usage{
SAFTE_model(dataset, bedtime = "23:00")
}
\arguments{
\item{dataset}{Dataset output from SAFTE_epoch_tbl}

\item{bedtime}{Goal or average bedtime of the subject. Should be formatted in 24hr time as "HH:MM". Default is set to"23:00" or 11PM}
}
\value{
This will output a table listing all 24 hour periods that do not have an instance of sleep.
}
\description{
This function takes the SAFTE_epoch_tbl output and runs it through the SAFTE model
}
\details{
The current algorithm is still being refactored for quicker calculations. The current
function can run 1+ minutes for final output. A progress bar should display (progress package)
}
