% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFTE_plot.R
\name{SAFTE_plot}
\alias{SAFTE_plot}
\title{SAFTE Plot}
\usage{
SAFTE_plot(
  dataset,
  title = "SAFTE Plot",
  plot_var = "Effectiveness",
  x_label = "Observation Day",
  y_label = plot_var,
  work = FALSE,
  test = FALSE,
  crewing = FALSE,
  settings = TRUE,
  print = TRUE
)
}
\arguments{
\item{dataset}{dataset output from SAFTE_model}

\item{title}{Name/Title to be displayed at the top of the plot. Defaults to "SAFTE Plot"}

\item{plot_var}{The variable to be plotted. Defaults to "Effectiveness". Variable should be name of column in " "}

\item{x_label}{The label for the x-axis (days of observation). Defaults to "Observation Day".}

\item{y_label}{The label for the y-axis. Defaults to the name of the plot_var}

\item{work}{Include work markers if in dataset. Defaults to FALSE,}

\item{test}{Include test markers if in dataset. Defaults to FALSE.}

\item{crewing}{Include crewing markers if in dataset. Defaults to FALSE.}

\item{settings}{Include a subtitle with subjectID. Defaults to TRUE.}

\item{print}{Print the plot after function run}
}
\value{
This will output an object that contains all of the information for the ggplot to be plotted.
It can also return a print of the plot if print is set to TRUE.
}
\description{
This function helps plot the fully realized data from the SAFTE model and combines
with Sleep, Work, Test, and Crewing markers for visualizing across a timeline.
}
\details{
Effectiveness and Reservoir_Percent are the most common variables to plot.

This funciton utilizes ggplot2 for most graphing functions
}
