% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAFTE_plot_compare_ln.R
\name{SAFTE_plot_compare_ln}
\alias{SAFTE_plot_compare_ln}
\title{SAFTE Plot Compare-Line Graph}
\usage{
SAFTE_plot_compare_ln(
  dataset,
  plot_var = "Effectiveness",
  compare_data,
  datetime_col,
  compare_col,
  title = "SAFTE Plot",
  x_label = "Observation Day",
  y_label_1 = plot_var,
  y_label_2 = compare_col,
  work = FALSE,
  test = FALSE,
  crewing = FALSE,
  settings = TRUE,
  print = TRUE
)
}
\arguments{
\item{dataset}{dataset output from SAFTE_model}

\item{plot_var}{The variable to be plotted. Defaults to "Effectiveness". Variable should be name of column in " "}

\item{compare_data}{dataset with data to be compared}

\item{datetime_col}{column name that has the datetime to be compared to the SAFTE model data}

\item{compare_col}{column name of the data to be compared to the SAFTE model data}

\item{title}{Name/Title to be displayed at the top of the plot. Defaults to "SAFTE Plot"}

\item{x_label}{The label for the x-axis (days of observation). Defaults to "Observation Day".}

\item{y_label_1}{The label for the y-axis of the SAFTE model. Defaults to the name of the plot_var}

\item{y_label_2}{The label for the y-axis of the second y-axis of the compare data. Defaults to the name of the compare_col}

\item{work}{Include work markers if in dataset. Defaults to FALSE,}

\item{test}{Include test markers if in dataset. Defaults to FALSE.}

\item{crewing}{Include crewing markers if in dataset. Defaults to FALSE.}

\item{settings}{Include a subtitle with subjectID. Defaults to TRUE.}

\item{print}{Print the plot after function run}
}
\value{
This will output an object that contains all of the information for the ggplot to be plotted.
It can also return a print of the plot if print is set to TRUE.

This plots a line graph of the compare data against a line of the SAFTE model.
}
\description{
This function helps plot the fully realized data from the SAFTE model and a dataset
from the same timeframe. It also combines with Sleep, Work, Test, and
Crewing markers for visualizing across a timeline.
}
\details{
Compare data must sync with the SAFTE model data by observation datetimes.

Effectiveness and Reservoir_Percent are the most common variables to plot.

This function utilizes ggplot2 for most graphing functions.

This is creates a line graph of the compare data.
}
